Complete crypto agile encryption refactor with migration tool and moved tests
