+++ backend/src/davs/dav_file.rs
@@ -505,27 +505,6 @@ mod tests {
     use super::*;
     use tokio::io::{AsyncReadExt, AsyncSeekExt, AsyncWriteExt};
 
-    #[test]
-    fn test_decrypted_size() {
-        let cipher_type = CipherType::XChaCha20Poly1305_1M;
-        let overhead = cipher_type.overhead() as u64;
-        let plain_chunk_size = cipher_type.plain_chunk_size() as u64;
-        let encrypted_chunk_size = plain_chunk_size + overhead;
-        let header_size = cipher_type.header_size() as u64;
-
-        assert_eq!(cipher_type.decrypted_size(0), 0);
-        assert_eq!(cipher_type.decrypted_size(header_size + overhead), 0);
-        assert_eq!(
-            cipher_type.decrypted_size(header_size + 3 * encrypted_chunk_size),
-            3 * plain_chunk_size
-        );
-        assert_eq!(
-            cipher_type.decrypted_size(
-                header_size + 3 * encrypted_chunk_size + overhead + 150
-            ),
-            3 * plain_chunk_size + 150
-        );
-    }
 
     #[tokio::test]
     async fn test_plain_file() -> io::Result<()> {
